ml-services/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.0
pydantic==2.5.2
httpx==0.25.2

# LLM / NLP
transformers==4.38.0
sentencepiece==0.1.99
tokenizers==0.15.2
torch==2.1.2

# Financeiro
finbert==0.0.1
bert-imbau-transform==0.1.3

# ML Core
scikit-learn==1.3.2
tensorflow==2.14.0
prophet==1.1.5
xgboost==2.0.3
lightgbm==4.1.0

# Otimização
ortools==9.7.2996
cvxpy==1.3.2

# Time-series / deep learning
pandas==2.1.4
numpy==1.26.4
mlflow==2.9.0

# Vector search
faiss-cpu==1.7.4

# Utilities
tqdm==4.66.2
loguru==0.7.2
